requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.2
requests==2.31.0
beautifulsoup4==4.12.3
plotly==5.18.0
python-dotenv==1.0.0
schedule==1.2.0